feat(*): starting style refactoring
