+++ src-ebleague/app/admin/admin-routing.module.ts
@@ -6,11 +6,14 @@ import { AdminCsgoComponent } from './admin-csgo/admin-csgo.component';
 import { AdminGuard } from '../../../src-common/guards/admin.guard';
 import { EventAdminGuard } from '../../../src-common/guards/eventadmin.guard';
 import { AdminSc2Component } from './admin-sc2/admin-sc2.component';
+import { AdminMainComponent } from './admin-main/admin-main.component';
 
 const routes: Routes = [
-  { path: '', component: AdminComponent, canActivate: [ AdminGuard ] },
-  { path: 'csgo', component: AdminCsgoComponent, canActivate: [ EventAdminGuard ] },
-  { path: 'sc2', component: AdminSc2Component, canActivate: [ EventAdminGuard ] }
+  { path: '', component: AdminComponent, children: [
+    { path: '', component: AdminMainComponent, canActivate: [ AdminGuard ] },
+    { path: 'csgo', component: AdminCsgoComponent, canActivate: [ EventAdminGuard ] },
+    { path: 'sc2', component: AdminSc2Component, canActivate: [ EventAdminGuard ] }
+  ] }
 ];
 
 @NgModule({
