+++ src-ebleague/app/admin/admin.component.scss
@@ -1,20 +1,81 @@
-:host {
-  .players-list {
-    max-height: 500px;
-    overflow: auto;
+@import "~@infragistics/igniteui-angular/lib/core/styles/themes/index";
+@import "../../style-vars.scss";
+
+:host ::ng-deep {
+  /* Grid */
+  $grid-theme: igx-grid-theme($content-background: $banner-background,
+                              $drop-area-background: $banner-background,
+                              $ghost-header-background: $banner-background,
+                              $group-row-selected-background: $background,
+                              $header-background: $background,
+                              $cell-editing-background: $background);
+  @include igx-css-vars($grid-theme);
+  /* END Grid */
+
+  /* Grid Toolbar */
+  $grid-toolbar-theme: igx-grid-toolbar-theme($background-color: $background);
+  @include igx-css-vars($grid-toolbar-theme);
+  /* END Grid Toolbar */
+
+  /* Paginator */
+  $paginator-theme: igx-paginator-theme($background-color: $background);
+  @include igx-css-vars($paginator-theme);
+  /* END Paginator */
+
+  /* Action strip */
+  $action-strip-theme: igx-action-strip-theme($actions-background: transparent);
+  @include igx-css-vars($action-strip-theme);
+  /* END Action strip */
+
+  .registration-container {
+    padding-bottom: 12px;
   }
 
-  .details-container {
-    width: 100%;
+  .new-group-card {
+    border: 1px dashed #aaa;
   }
 
-  .column-center {
+  .new-group-card:hover {
+    border: 1px dashed #fff;
+  }
+
+  .new-group-body {
+    min-height: 350px;
     display: flex;
-    flex-flow: column;
+    width: 100%;
     align-items: center;
+    justify-content: center;
   }
 
-  .content-container {
-    min-height: auto;
+  .new-group-body igx-icon {
+    width: 120px;
+    height: 120px;
+    font-size: 90pt;
+  }
+
+  .group-list {
+    min-height: 350px;
+  }
+
+  .new-group-body igx-icon:hover {
+    color: #fff;
+  }
+
+  .group-card-content {
+    padding: 0;
+  }
+
+  .action-icon-button {
+    margin: 0 4px;
+  }
+
+  .action-icon-button:hover {
+    color: $secondary;
+  }
+
+  .action-buttons {
+    display: flex;
+    align-items: center;
+    padding-right: 4px;
   }
 }
