+++ src-ebleague/app/admin/admin.component.spec.ts
@@ -1,21 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AdminComponent } from './admin.component';
-import { IgxChipsModule,
-  IgxInputGroupModule,
-  IgxButtonModule,
-  IgxAvatarModule,
-  IgxIconModule,
-  IgxDatePickerModule,
-  IgxGridModule,
-  IgxActionStripModule,
-  IgxCheckboxModule} from '@infragistics/igniteui-angular';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ServiceWorkerModule } from '@angular/service-worker';
-import { RouterTestingModule } from '@angular/router/testing';
-import { FormsModule } from '@angular/forms';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { SizeNamePipe } from '../pipes/size-name.pipe';
 
 describe('AdminComponent', () => {
   let component: AdminComponent;
@@ -24,24 +9,7 @@ describe('AdminComponent', () => {
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [
-        AdminComponent,
-        SizeNamePipe
-      ],
-      imports: [
-        FormsModule,
-        HttpClientTestingModule,
-        NoopAnimationsModule,
-        RouterTestingModule,
-        ServiceWorkerModule.register('', {enabled: false}),
-        IgxChipsModule,
-        IgxInputGroupModule,
-        IgxButtonModule,
-        IgxAvatarModule,
-        IgxIconModule,
-        IgxDatePickerModule,
-        IgxGridModule,
-        IgxActionStripModule,
-        IgxCheckboxModule
+        AdminComponent
       ]
     })
     .compileComponents();
