+++ src-ebleague/app/tournaments/tournament-csgo/tournament-csgo.component.scss
@@ -1,4 +1,22 @@
+@import "~@infragistics/igniteui-angular/lib/core/styles/themes/index";
+@import "../../../style-vars.scss";
+
 :host {
+  /* Grid */
+  $grid-theme: igx-grid-theme($content-background: $banner-background,
+                              $drop-area-background: $banner-background,
+                              $ghost-header-background: $banner-background,
+                              $group-row-selected-background: $background,
+                              $header-background: $background,
+                              $cell-editing-background: $background);
+  @include igx-css-vars($grid-theme);
+  /* END Grid */
+
+  /* Grid Toolbar */
+  $grid-toolbar-theme: igx-grid-toolbar-theme($background-color: $background);
+  @include igx-css-vars($grid-toolbar-theme);
+  /* END Grid Toolbar */
+
   .content-container.two-column {
     width: 100%;
   }
