+++ src-ebleague/app/tournaments/tournament-sc2/tournament-sc2.component.scss
@@ -1,23 +1,43 @@
-@media only screen and (min-width: 1280px) {
-  .content-container.four-column {
-    width: 100%;
-    grid-template-columns: minmax(20%, 1fr) minmax(20%, 1fr) minmax(20%, 1fr) minmax(20%, 1fr);
-  }
+@import "~@infragistics/igniteui-angular/lib/core/styles/themes/index";
+@import "../../../style-vars.scss";
 
-  .group-columns {
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-row-gap: 16px;
-    padding: 48px 0;
-  }
+:host {
+  /* Grid */
+  $grid-theme: igx-grid-theme($content-background: $banner-background,
+                              $drop-area-background: $banner-background,
+                              $ghost-header-background: $banner-background,
+                              $group-row-selected-background: $background,
+                              $header-background: $background,
+                              $cell-editing-background: $background);
+  @include igx-css-vars($grid-theme);
+  /* END Grid */
 
-  @media only screen and (min-width: 1921px) {
+  /* Grid Toolbar */
+  $grid-toolbar-theme: igx-grid-toolbar-theme($background-color: $background);
+  @include igx-css-vars($grid-toolbar-theme);
+  /* END Grid Toolbar */
+
+  @media only screen and (min-width: 1280px) {
+    .content-container.four-column {
+      width: 100%;
+      grid-template-columns: minmax(20%, 1fr) minmax(20%, 1fr) minmax(20%, 1fr) minmax(20%, 1fr);
+    }
+
     .group-columns {
       display: grid;
-      grid-template-columns: minmax(45%, 1fr) minmax(45%, 1fr);
-      grid-column-gap: 48px;
+      grid-template-columns: 1fr;
       grid-row-gap: 16px;
       padding: 48px 0;
+    }
+
+    @media only screen and (min-width: 1921px) {
+      .group-columns {
+        display: grid;
+        grid-template-columns: minmax(45%, 1fr) minmax(45%, 1fr);
+        grid-column-gap: 48px;
+        grid-row-gap: 16px;
+        padding: 48px 0;
+      }
     }
   }
 }
