+++ src-ebleague/styles.scss
@@ -148,31 +148,6 @@ $time-picker-theme: igx-time-picker-theme($background-color: $background);
 @include igx-css-vars($time-picker-theme);
 /* END Time Picker */
 
-/* Grid */
-$grid-theme: igx-grid-theme($content-background: $banner-background,
-                            $drop-area-background: $banner-background,
-                            $ghost-header-background: $banner-background,
-                            $group-row-selected-background: $background,
-                            $header-background: $background,
-                            $cell-editing-background: $background);
-@include igx-css-vars($grid-theme);
-/* END Grid */
-
-/* Grid Toolbar */
-$grid-toolbar-theme: igx-grid-toolbar-theme($background-color: $background);
-@include igx-css-vars($grid-toolbar-theme);
-/* END Grid Toolbar */
-
-/* Paginator */
-$paginator-theme: igx-paginator-theme($background-color: $background);
-@include igx-css-vars($paginator-theme);
-/* END Paginator */
-
-/* Action strip */
-$action-strip-theme: igx-action-strip-theme($actions-background: transparent);
-@include igx-css-vars($action-strip-theme);
-/* END Action strip */
-
 @import "../src-common/styles-overrides.scss";
 
 /* HOME */
@@ -202,17 +177,9 @@ $action-strip-theme: igx-action-strip-theme($actions-background: transparent);
 .slide-info-column p {
   font-size: 1rem;
 }
-
-.twitch-holder {
-  text-align: center;
-}
 /* END HOME */
 
 /* Registrations */
-.registration-container {
-  padding-bottom: 12px;
-}
-
 .registration-slot {
   position: relative;
   display: flex;
@@ -233,77 +200,13 @@ $action-strip-theme: igx-action-strip-theme($actions-background: transparent);
   position: absolute;
   left: 86px;
 }
-
-.empty-group-list {
-  height: 300px;
-  display: flex;
-  align-items: center;
-}
 /* END Registrations */
 
-/* Admin */
-.new-group-card {
-  border: 1px dashed #aaa;
-}
-
-.new-group-card:hover {
-  border: 1px dashed #fff;
-}
-
-.new-group-body {
-  min-height: 350px;
-  display: flex;
-  width: 100%;
-  align-items: center;
-  justify-content: center;
-}
-
-.new-group-body igx-icon {
-  width: 120px;
-  height: 120px;
-  font-size: 90pt;
-}
-
-.group-list {
-  min-height: 350px;
-}
-
-.new-group-body igx-icon:hover {
-  color: #fff;
-}
-
-.group-card-content {
-  padding: 0;
-}
-
-.action-icon-button {
-  margin: 0 4px;
-}
-
-.action-icon-button:hover {
-  color: $secondary;
-}
-
-.action-buttons {
-  display: flex;
-  align-items: center;
-  padding-right: 4px;
-}
-/* END Admin */
-
 /* SCHEDULE */
-.card-wrapper.igx-tabs__group {
-  display: grid;
-}
-
 .participants-container {
   display: flex;
   flex-wrap: wrap;
   padding: 16px;
-}
-
-igx-calendar {
-  height: fit-content;
 }
 
 .match-actions > * {
